Relocate types file and add Request type so that it can include custom properties
