+++ src/lib/routes/order.ts
@@ -1,6 +1,6 @@
 import { Router } from 'express';
 import { getOrders, getOrder, addOrder, updateOrder } from '../models/order';
-import type { TechComponentOrder } from '../types';
+import type { TechComponentOrder } from '../types/types';
 
 const router = Router();
