+++ src/lib/types/express/index.d.ts
@@ -1,0 +1,5 @@
+declare namespace Express {
+  interface Request {
+    email?: string;
+  }
+}
